Refactor and update CI for encoder-decoders (#742)
